translateblogs/requirements.txt
pypinyin
openai
python-frontmatter 
pyyaml 
tqdm